add xbuild for srvice.

+++ src/include/stdint.h
@@ -1,8 +1,8 @@
 #ifndef _XBOOK_STDINT_H
 #define _XBOOK_STDINT_H
 
-typedef unsigned long   uint64_t;
-typedef signed long     int64_t;
+typedef unsigned long long uint64_t;
+typedef signed long long int64_t;
 typedef unsigned int    uint32_t;
 typedef signed int      int32_t;
 typedef unsigned short  uint16_t;
@@ -10,8 +10,8 @@ typedef signed short    int16_t;
 typedef unsigned char   uint8_t;
 typedef signed char     int8_t;
 
-typedef unsigned long   uint64;
-typedef signed long     int64;
+typedef unsigned long long uint64;
+typedef signed long long int64;
 typedef unsigned int    uint32;
 typedef signed int      int32;
 typedef unsigned short  uint16;
@@ -19,8 +19,8 @@ typedef signed short    int16;
 typedef unsigned char   uint8;
 typedef signed char     int8;
 
-typedef unsigned long   u64;
-typedef signed long     s64;
+typedef unsigned long long u64;
+typedef signed long long s64;
 typedef unsigned int    u32;
 typedef signed int      s32;
 typedef unsigned short  u16;
